displayLineScanLocation bluefish compatibility
Turns out there are some differences between bluefish and thing 1, so I had to make a few rig checks and modifications. The function doesn't work as well on bluefish for some reason. It seems that the scaling in the y axis (and possibily x) is slightly off. This will need to be addressed at a later date