added gsat functionality + other improvements
There is now the ability to enter gsat for a linescan set, and the plotting functions now check for gstat in the expParams.

I also added autoscale capabilities for the yaxis to facilitate changing between df/f and dG/R.

Finally, I enabled dynamically editing either single or all of the spike times and gsat values for a given linescan set. If the plot is set to group plot, values for all scans will be changed, whereas in single or sweep plots only the individual scan's values will be changed.